crypto-tracking-function/requirements.txt
azure-functions
azure-data-tables
azure-core
